Gather Data From Scopus

This function gather data from Scopus database. You just need to input the journal names that you want to do a research with, and api key and token that you got from Scopus, then you can get the data.

The data will include these columns:

"AU" -- Authors Names;

"TI" -- Document Title;

"AID" -- Author's ID;

"PY" -- Publication Year;

"SO" -- Journal Name (or Source);

"TC" -- Times Cited;

"C1" -- Affiliations;

"AB" -- Abstract;

"DE" -- Keywords;

"ID" -- Keywords associated by SCOPUS or WoS database;

"CR" -- Cited References;

"RP" -- Corresponding Author Affiliation;

"JI" -- ISO Source Abbreviation;

"DT" -- Document Type;

"DB" -- Bibliographic Database;

"UT" -- Unique Article Identifier;

gather_scopus(
  journals,
  api_key,
  insttoken,
  abbreviate = FALSE,
  year_greater = NULL,
  year_less = NULL
)

Examples

gather_scopus()

#> Error in gather_scopus(): argument "journals" is missing, with no default